Add spy detection feature
Spies are clients hidden from other players on the server.

The detection feature relies on finding clientnum irregularities when a 
new player joins the server.

Spies whose clientnum is greater than that of the connected player 
cannot be detected.

False positives can occur when a player reconnects to the server.

To enable: /detectspies 1